requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0
PyCryptodome>=3.15
Flask-Login>=0.6.0
Flask-Bcrypt>=1.0.0
twilio>=8.0.0
python-dotenv>=1.0
